+++ core/device.h
@@ -17,6 +17,7 @@ OIDN_NAMESPACE_BEGIN
   class Subdevice;
   class Engine;
   class Buffer;
+  class Semaphore;
   class Filter;
 
   class PhysicalDevice : public RefCount
@@ -77,10 +78,10 @@ OIDN_NAMESPACE_BEGIN
     Ref<Buffer> newUserBuffer(void* ptr, size_t byteSize);
     Ref<Buffer> newNativeUserBuffer(void* handle);
 
-    Ref<Buffer> newExternalUserBuffer(ExternalMemoryTypeFlag fdType,
+    Ref<Buffer> newExternalUserBuffer(ExternalMemoryTypeFlags fdType,
                                       int fd, size_t byteSize);
 
-    Ref<Buffer> newExternalUserBuffer(ExternalMemoryTypeFlag handleType,
+    Ref<Buffer> newExternalUserBuffer(ExternalMemoryTypeFlags handleType,
                                       void* handle, const void* name, size_t byteSize);
 
     // Filter
@@ -109,6 +110,22 @@ OIDN_NAMESPACE_BEGIN
     bool isManagedMemorySupported() const { return managedMemorySupported; }
     ExternalMemoryTypeFlags getExternalMemoryTypes() const { return externalMemoryTypes; }
     void trimScratch();
+
+    // Semaphore
+    Ref<Semaphore> newExternalSemaphore(ExternalSemaphoreTypeFlags fdType, int fd);
+    Ref<Semaphore> newExternalSemaphore(ExternalSemaphoreTypeFlags handleType,
+                                        void* handle, const void* name);
+
+    void submitSignalSemaphores(Semaphore* const* semaphores,
+                                const uint64_t* values,
+                                int numSemaphores);
+
+    void submitWaitSemaphores(Semaphore* const* semaphores,
+                              const uint64_t* values,
+                              const uint32_t* timeoutsMs,
+                              int numSemaphores);
+
+    ExternalSemaphoreTypeFlags getExternalSemaphoreTypes() const { return externalSemaphoreTypes; }
 
     // Executes operations on the device, making sure to wait/flush and release temporary
     // allocations (e.g. from ObjC) at the end, even if an exception is thrown
@@ -146,6 +163,8 @@ OIDN_NAMESPACE_BEGIN
     bool systemMemorySupported  = false;
     bool managedMemorySupported = false;
     ExternalMemoryTypeFlags externalMemoryTypes;
+
+    ExternalSemaphoreTypeFlags externalSemaphoreTypes;
 
     // State
     bool dirty = true;