+++ devices/hip/hip_device.cpp
@@ -200,9 +200,22 @@ OIDN_NAMESPACE_BEGIN
                           ExternalMemoryTypeFlag::D3D11ResourceKMT |
                           ExternalMemoryTypeFlag::D3D12Heap |
                           ExternalMemoryTypeFlag::D3D12Resource;
+
+    externalSemaphoreTypes = ExternalSemaphoreTypeFlag::OpaqueWin32 |
+                             ExternalSemaphoreTypeFlag::OpaqueWin32KMT |
+                             ExternalSemaphoreTypeFlag::D3D11Fence |
+                             ExternalSemaphoreTypeFlag::D3D12Fence |
+                             ExternalSemaphoreTypeFlag::KeyedMutex |
+                             ExternalSemaphoreTypeFlag::KeyedMutexKMT |
+                             ExternalSemaphoreTypeFlag::TimelineSemaphoreWin32;
   #else
     externalMemoryTypes = ExternalMemoryTypeFlag::OpaqueFD;
+
+    // Semaphores not supported by HIP on Linux yet
+    externalSemaphoreTypes = ExternalSemaphoreTypeFlag::None;
   #endif
+
+    externalMemoryTypes |= ExternalMemoryTypeFlag::Dedicated;
 
     subdevices.emplace_back(new Subdevice(std::unique_ptr<Engine>(new HIPEngine(this, stream))));
   }